requirements.txt
Flask==2.3.3
requests==2.31.0
twilio==8.10.0
schedule==1.2.0
python-dotenv==1.0.0